services/process-manager/requirements.txt
fastapi
uvicorn
python-dotenv
psutil   # для работы с процессами
pyyaml